api services directory restructure
