+++ src/services/api/endpoint.js
@@ -1,0 +1,8 @@
+export default {
+  MOVIES: '/movies',
+  GENRES: '/movie_genres',
+  SEANCES: '/seances',
+  USER: '/user',
+  LOGIN: '/login',
+  REGISTER: '/register',
+};
